gallium/auxiliary: Use 32-bit element indices.
not entirely sure this is correct.  some comments
suggest this widening of sizes

but other comments say vsplit should guarantee
small batches of vertices and more complex vsplit
changes are needed for the geometry/tesselation
cases where large numbers of vertices are emitted

+++ src/gallium/auxiliary/draw/draw_gs.c
@@ -561,7 +561,7 @@ static void gs_tri_adj(struct draw_geometry_shader *shader,
 
 
 #define FUNC         gs_run_elts
-#define LOCAL_VARS   const ushort *elts = input_prims->elts;
+#define LOCAL_VARS   const unsigned *elts = input_prims->elts;
 #define GET_ELT(idx) (elts[idx])
 #include "draw_gs_tmp.h"
